Recovered from Eclipse.
git-svn-id: https://subversion.assembla.com/svn/cmosher@7508 884d43a4-ffbe-47e9-90ca-fd8c37d7eb32

+++ diehard/src/edu/fsu/stat/geo/diehard/tests/RandomnessTest.java
@@ -20,7 +20,7 @@ public abstract class RandomnessTest
     {
         int intr = rng.nextInt();
 
-        long r = (long)intr;
+        long r = intr;
 
         if (intr < 0)
             r += 1L<<32;
